requirements.txt
praw
b2sdk
python-dotenv
flask
pandas
emoji
pyyaml